requirements.txt
# Base
Pillow==8.3.2
pytz==2021.3
python-decouple==3.5

# Django
Django==3.2.8

# Django REST Framework
djangorestframework==3.12.4
django-filter==21.1

# Auth
PyJWT==2.2.0
django-rest-framework-social-oauth2==1.1.0

# Enviroment
django_environ==0.7.0

# Postgres
psycopg2==2.9.1

# Celery
celery==5.1.2
redis==3.5.3

# Tests
coverage==6.1.1